Add support for hardlinking or symlinking files.
This change adds the new --file_strategy flag to build_pack.py which
adds support for hardlinking or symlinking files instead of the default
behavior of copying files.

This should be useful for users on space constrained disks, and
additionally should be faster than copying files.

The default is left at copy however, since you can't trivially copy a
symlink tree to an SD card and expect it to work, and hardlinks can't be
created which cross filesystem boundaries, so the user is forced to make
the conscious choice to use hardlinks.