fix for the list of missing files (with support for zipped archived)
SMDB database entries are stored in memory using a dictionary with
SHA256 values as key. When a file is matched to a database
entry (based on its SHA256 value), the file is processed and the
database entry is deleted from the dict. Finally, all the database
entries remaining at the end of the process are considered as missing.

Support for zipped folders required to store SMDB database entries
using CRC32 values as key, in addition to using SHA256
values. However, only one of each key is deleted from the database
during the build process, resulting in an inflated list of missing
files.

Observed files will have either (or both) the SHA256 or the CRC32
entry deleted from the database. Missing files will still have both
entries present in the database at the end of the build process. So,
the idea is to search for filenames occuring twice to establish a
valid list of missing files.

My tests show that it works as expected when missing normal and zipped
archives.

An alternative would be to store SHA256 and CRC32 in two separate
databases, but that would also require some extra-process to work with
mixed input (normal and zipped).