remove dest file if it exists to avoid a FileExistsError
When trying to update an existing folder, if the --skip_existing
option is not activated, the script stops with a FileExistsError. In
practice, this means that files in an existing pack cannot be
overwritten.

Now, when copying a file from its source to its destination, if a file
with the same name already exists at the destination and the
--skip_existing is set to false, destination is removed before the
copy.