SegaCD (trimmed redump): remove redundant entries from A-M
I believe the goal of this SMDB was to be a 1G1R set for Sega CD (only one version of each game, with preference to USA, Japan, Europe in that order). This commit removes some duplicates that were missed due to naming differences across regions. It also removes a non-game entry (Fresh Cleaner) which is "software that cleans the system's disk drive lens while playing an animation"