Neo Geo - Fix formatting, move test rom to tools folder, and add kof98ae
To standardize the use of Tools and Service Test Carts folder like other DB entries, and because there are undoubtedly other test roms out there, I've moved the monitor test rom into an appropriate test folder. Also htgdb-packs on archive had kof98ae, but it was missing from the db for some reason. also the naming was inconsistent (The King of Fighters vs King of Fighters, The which is more appropriate, mixed).