ignore the new size column when parsing DB file
TORQUENDB described the issue:

build_pack.py only uses the CRC hash when checking zipped files
against the database, instead of the SHA256 it uses for normal
files. However, when reading the DB files, it only splits by 4 tabs,
and if a DB file has the new size column (e,g. SNES, Wonderswan), the
hash_crc variable contains the CRC value, a tab, and the file
size. This makes it impossible to use zipped ROMs with these DB files
because the hash comparison always fails.