SNES: add Rendering Ranger R2 Bugfix hack
This hack for Rendering Ranger R2 allows the game to boot on the SNES Mini and on the current version of the MiSTer SNES core. Rendering Ranger R2 has some inherent bugs in its original design that makes it too sensitive to the varying clock speed of the sound chips between variation in original hardware.